+++ java/greeter/types/src/main/java/types/Internal/InternalOrderFindCallback.java
@@ -4,15 +4,15 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class InternalOrderFindCallback {
 	@JsonProperty("summed_cost")
-	private Integer summed_cost;
+	private Float summed_cost;
 
 	public InternalOrderFindCallback() {}
 
-	public InternalOrderFindCallback(int summed_cost) {
+	public InternalOrderFindCallback(Float summed_cost) {
 		this.summed_cost = summed_cost;
 	}
 
-	public Integer getSummed_cost() {
+	public Float getSummed_cost() {
 		return summed_cost;
 	}
 }
